implement PR suggestions
